requirements.txt
uWSGI==2.0.22
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Excel==0.0.7
PyMySQL==1.1.0
requests==2.31.0
cryptography==41.0.3
python-dateutil==2.8.2

